feat(smart-mode): configuration option for smart-mode base revision
